[localstorage] Customizable localstorage keys
In the case where theia instances would be behind a proxy, the pathname becomes irrelevant. It could point to anything,
could be changed, but we don't want the users to loose the configuration each time.

This is why this patch adds the possibility to define an ID for theia as a command line argument, so that we can identify
the instance on which we run.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>