make `MessageService` public API
`MessageService` is now a symbol+interface, it previously was a class.

When making abstractions public API, we need to separate API from
implementation, this is what this commit does by separating both.

Renamed and updated `MessageClient` to `MessageServer`, as it is not a
client but rather a server to be used over RPC.